requirements.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
prompt-toolkit==1.0.14
pycryptodome==3.10.1
Pygments==2.8.1
PyInquirer==1.0.3
pywin32-ctypes==0.2.0
regex==2021.4.4
six==1.15.0
wcwidth==0.2.5
